requirements.txt
openai>=1.50.0
python-telegram-bot>=21.0
python-dotenv>=1.0.1
